feat(router): add user and userinfo's router
main.py change to app.py and add router file
