+++ backend/database.py
@@ -13,7 +13,7 @@ SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 Base = declarative_base()
 
 
-def get_db():
+def db():
     db = SessionLocal()
     try:
         yield db
